CRUD-metoder til oprettelse af ny film - backend skulle gerne være færdig. Herunder service, repo implementeret.
