+++ src/main/java/stud/kea/dk/biografbackend/movie/service/MovieService.java
@@ -1,0 +1,35 @@
+package stud.kea.dk.biografbackend.movie.service;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import stud.kea.dk.biografbackend.movie.model.MovieModel;
+import stud.kea.dk.biografbackend.movie.repository.MovieRepository;
+
+import java.util.List;
+
+@Service
+public class MovieService {
+
+    @Autowired
+    private MovieRepository movieRepository;
+
+    public MovieModel createMovie(MovieModel movie) {
+        return movieRepository.save(movie);
+    }
+
+    public List<MovieModel> getAllMovies() {
+        return movieRepository.findAll();
+    }
+
+    public MovieModel getMovieById(int id) {
+        return movieRepository.findById(id).orElse(null);
+    }
+
+    public MovieModel updateMovie(MovieModel movie) {
+        return movieRepository.save(movie);
+    }
+
+    public void deleteMovie(int id) {
+        movieRepository.deleteById(id);
+    }
+}
